Added tests folder for integration tests
