+++ src/plonk.rs
@@ -121,8 +121,12 @@ impl<C: HaloCurve> Circuit<C> {
         // Commit to Z.
         let plonk_z_coeffs =
             ifft_with_precomputation_power_of_2(&plonk_z_points_n, &self.fft_precomputation_n);
-        let c_plonk_z =
-            PolynomialCommitment::coeffs_to_commitment(&plonk_z_coeffs, &self.pedersen_g_msm_precomputation, self.pedersen_h, blinding_commitments);
+        let c_plonk_z = PolynomialCommitment::coeffs_to_commitment(
+            &plonk_z_coeffs,
+            &self.pedersen_g_msm_precomputation,
+            self.pedersen_h,
+            blinding_commitments,
+        );
 
         // Generate a random alpha from the transcript.
         challenger.observe_affine_point(c_plonk_z.to_affine());
@@ -173,8 +177,12 @@ impl<C: HaloCurve> Circuit<C> {
             .collect();
 
         // Commit to the quotient polynomial.
-        let c_plonk_t =
-            PolynomialCommitment::coeffs_vec_to_commitments(&plonk_t_coeff_chunks, &self.pedersen_g_msm_precomputation, self.pedersen_h, blinding_commitments);
+        let c_plonk_t = PolynomialCommitment::coeffs_vec_to_commitments(
+            &plonk_t_coeff_chunks,
+            &self.pedersen_g_msm_precomputation,
+            self.pedersen_h,
+            blinding_commitments,
+        );
 
         // Generate a random zeta from the transcript.
         challenger
@@ -660,150 +668,5 @@ impl<C: HaloCurve> Circuit<C> {
             z1,
             z2,
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{
-        verify_proof_circuit, CircuitBuilder, Curve, Field, PartialWitness, Tweedledee, Tweedledum,
-        NUM_WIRES,
-    };
-    use std::time::Instant;
-
-    #[test]
-    fn test_generate_proof_trivial() {
-        let mut builder = CircuitBuilder::<Tweedledee>::new(128);
-        let t = builder.constant_wire(<Tweedledee as Curve>::ScalarField::ZERO);
-        builder.assert_zero(t);
-        let mut partial_witness = PartialWitness::new();
-        partial_witness.set_target(t, <Tweedledee as Curve>::ScalarField::ZERO);
-        let circuit = builder.build();
-        let witness = circuit.generate_witness(partial_witness);
-        let proof = circuit.generate_proof::<Tweedledum>(witness, true).unwrap();
-        assert!(verify_proof_circuit::<Tweedledee, Tweedledum>(&[], &proof, &circuit,).is_ok());
-    }
-
-    #[test]
-    fn test_generate_proof_sum() {
-        let mut builder = CircuitBuilder::<Tweedledee>::new(128);
-        let t1 = builder.constant_wire(<Tweedledee as Curve>::ScalarField::ZERO);
-        let t2 = builder.constant_wire(<Tweedledee as Curve>::ScalarField::ZERO);
-        let s = builder.add(t1, t2);
-        builder.assert_zero(s);
-        let mut partial_witness = PartialWitness::new();
-        partial_witness.set_target(t1, <Tweedledee as Curve>::ScalarField::ZERO);
-        partial_witness.set_target(t2, <Tweedledee as Curve>::ScalarField::ZERO);
-        let circuit = builder.build();
-        let witness = circuit.generate_witness(partial_witness);
-        let proof = circuit.generate_proof::<Tweedledum>(witness, true).unwrap();
-        assert!(verify_proof_circuit::<Tweedledee, Tweedledum>(&[], &proof, &circuit,).is_ok());
-    }
-
-    #[test]
-    fn test_generate_proof_sum_big() {
-        let now = Instant::now();
-        let mut builder = CircuitBuilder::<Tweedledee>::new(128);
-        let ts = (0..10_000)
-            .map(|i| {
-                builder.constant_wire(<Tweedledee as Curve>::ScalarField::from_canonical_usize(i))
-            })
-            .collect::<Vec<_>>();
-        let s = builder.add_many(&ts);
-        let x = builder.add_virtual_target();
-        let z = builder.sub(s, x);
-        builder.assert_zero(z);
-        let mut partial_witness = PartialWitness::new();
-        partial_witness.set_target(
-            x,
-            <Tweedledee as Curve>::ScalarField::from_canonical_usize((10_000 * 9_999) / 2),
-        );
-        dbg!(now.elapsed());
-        let circuit = builder.build();
-        dbg!(now.elapsed());
-        let witness = circuit.generate_witness(partial_witness);
-        dbg!(now.elapsed());
-        let proof = circuit.generate_proof::<Tweedledum>(witness, true).unwrap();
-        dbg!(now.elapsed());
-        assert!(verify_proof_circuit::<Tweedledee, Tweedledum>(&[], &proof, &circuit,).is_ok());
-        dbg!(now.elapsed());
-    }
-
-    #[test]
-    fn test_generate_proof_quadratic() {
-        let mut builder = CircuitBuilder::<Tweedledee>::new(128);
-        let one = builder.one_wire();
-        let t = builder.add_virtual_target();
-        let t_sq = builder.square(t);
-        let quad = builder.add_many(&[one, t, t_sq]);
-        let seven =
-            builder.constant_wire(<Tweedledee as Curve>::ScalarField::from_canonical_usize(7));
-        let res = builder.sub(quad, seven);
-        builder.assert_zero(res);
-        let mut partial_witness = PartialWitness::new();
-        partial_witness.set_target(t, <Tweedledee as Curve>::ScalarField::TWO);
-        let circuit = builder.build();
-        let witness = circuit.generate_witness(partial_witness);
-        let proof = circuit.generate_proof::<Tweedledum>(witness, true).unwrap();
-        assert!(verify_proof_circuit::<Tweedledee, Tweedledum>(&[], &proof, &circuit,).is_ok());
-    }
-
-    #[test]
-    fn test_generate_proof_public_input1() {
-        // Set public inputs pi1 = 2 and check that pi1 - 2 == 0.
-        let mut builder = CircuitBuilder::<Tweedledee>::new(128);
-        let pi = builder.stage_public_input();
-        builder.route_public_inputs();
-        let t1 = pi.routable_target();
-        let t2 = builder.constant_wire(<Tweedledee as Curve>::ScalarField::TWO);
-        let t3 = builder.sub(t1, t2);
-        builder.assert_zero(t3);
-        let mut partial_witness = PartialWitness::new();
-        partial_witness.set_target(t1, <Tweedledee as Curve>::ScalarField::TWO);
-        let circuit = builder.build();
-        let witness = circuit.generate_witness(partial_witness);
-        let proof = circuit.generate_proof::<Tweedledum>(witness, true).unwrap();
-        // Check that the public input is set correctly in the proof.
-        assert_eq!(
-            proof.o_public_inputs[0].o_wires[0],
-            <Tweedledee as Curve>::ScalarField::TWO
-        );
-        assert!(verify_proof_circuit::<Tweedledee, Tweedledum>(
-            &[<Tweedledee as Curve>::ScalarField::TWO],
-            &proof,
-            &circuit,
-        )
-        .is_ok());
-    }
-
-    #[test]
-    #[ignore]
-    fn test_generate_proof_public_input2() {
-        // Set many random public inputs
-        let mut builder = CircuitBuilder::<Tweedledee>::new(128);
-        let pis = (0..200)
-            .map(|_| builder.stage_public_input())
-            .collect::<Vec<_>>();
-        builder.route_public_inputs();
-        let tis = pis.iter().map(|p| p.original_wire()).collect::<Vec<_>>();
-        let mut partial_witness = PartialWitness::new();
-        let values = tis
-            .iter()
-            .map(|&_t| <Tweedledee as Curve>::ScalarField::rand())
-            .collect::<Vec<_>>();
-        tis.iter().zip(values.iter()).for_each(|(&t, &v)| {
-            partial_witness.set_wire(t, v);
-        });
-        let circuit = builder.build();
-        let witness = circuit.generate_witness(partial_witness);
-        let proof = circuit.generate_proof::<Tweedledum>(witness, true).unwrap();
-        // Check that the public inputs are set correctly in the proof.
-        values.iter().enumerate().for_each(|(i, &v)| {
-            assert_eq!(
-                v,
-                proof.o_public_inputs[i / NUM_WIRES].o_wires[i % NUM_WIRES],
-            )
-        });
-        assert!(verify_proof_circuit::<Tweedledee, Tweedledum>(&values, &proof, &circuit,).is_ok());
     }
 }
